refactored: LoginView to Profile view file
