+++ Profile/urls.py
@@ -5,7 +5,6 @@ from Profile import views
 app_name = 'Profile'
 
 urlpatterns = [
-    path('profiles/', views.ProfileView.as_view({'get': 'retrieve'}), name='profile_view'),
-    path('profiles/<int:pk>/', views.ProfileView.as_view({'get': 'retrieve'}), name='profile_by_id'),
+    path('login/', views.LoginView.as_view({'post': 'retrieve'}), name='login_view'),
     path('profiles/change_password/', views.ChangePassword.as_view(), name='change_password'),
 ]
