+++ property/urls.py
@@ -19,7 +19,7 @@ from django.urls import path, include
 from rest_framework import routers
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
 
-from .views import HomeView, LoginView
+from .views import HomeView
 from .settings import LOCALHOST
 
 router = routers.DefaultRouter()
@@ -31,13 +31,8 @@ urlpatterns = [
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path('api-auth/', include('rest_framework.urls')),
-    # path('dj-rest-auth/', include('dj_rest_auth.urls')),
-    path('dj-rest-auth/login/', LoginView.as_view({'post': 'retrieve'}), name='rest_auth'),
     path('dj-rest-auth/registration/', include('dj_rest_auth.registration.urls')),
 
-    path('', include('Profile.urls')),
+    path('profiles/', include('Profile.urls')),
     path('stocks/', include('stocks.urls')),
 ]
-
-# if LOCALHOST == 'True':
-#     urlpatterns.append(path("__reload__/", include("django_browser_reload.urls")))
